LPS-179353 Extract ScreenNavigationEntry methods from CommerceTaxMethodFixedRatesScreenNavigationCategory to CommerceTaxMethodFixedRatesScreenNavigationEntry, extends CommerceTaxMethodFixedRatesScreenNavigationCategory in it and return getCategoryKey in getEntryKey method
