+++ modules/apps/commerce/commerce-tax-engine-fixed-web/src/main/java/com/liferay/commerce/tax/engine/fixed/web/internal/frontend/taglib/servlet/taglib/CommerceTaxMethodFixedRatesScreenNavigationCategory.java
@@ -15,35 +15,12 @@
 package com.liferay.commerce.tax.engine.fixed.web.internal.frontend.taglib.servlet.taglib;
 
 import com.liferay.commerce.constants.CommerceTaxScreenNavigationConstants;
-import com.liferay.commerce.currency.service.CommerceCurrencyLocalService;
-import com.liferay.commerce.percentage.PercentageFormatter;
-import com.liferay.commerce.product.model.CommerceChannel;
-import com.liferay.commerce.product.service.CPTaxCategoryService;
-import com.liferay.commerce.product.service.CommerceChannelLocalService;
-import com.liferay.commerce.tax.engine.fixed.service.CommerceTaxFixedRateService;
-import com.liferay.commerce.tax.engine.fixed.web.internal.display.context.CommerceTaxFixedRatesDisplayContext;
-import com.liferay.commerce.tax.model.CommerceTaxMethod;
-import com.liferay.commerce.tax.service.CommerceTaxMethodService;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationCategory;
-import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
-import com.liferay.frontend.taglib.servlet.taglib.util.JSPRenderer;
 import com.liferay.portal.kernel.language.Language;
-import com.liferay.portal.kernel.model.User;
-import com.liferay.portal.kernel.security.permission.resource.ModelResourcePermission;
-import com.liferay.portal.kernel.util.JavaConstants;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
-import com.liferay.portal.kernel.util.WebKeys;
-
-import java.io.IOException;
 
 import java.util.Locale;
 import java.util.ResourceBundle;
-
-import javax.portlet.RenderRequest;
-
-import javax.servlet.ServletContext;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -52,15 +29,11 @@ import org.osgi.service.component.annotations.Reference;
  * @author Alessio Antonio Rendina
  */
 @Component(
-	property = {
-		"screen.navigation.category.order:Integer=20",
-		"screen.navigation.entry.order:Integer=10"
-	},
-	service = {ScreenNavigationCategory.class, ScreenNavigationEntry.class}
+	property = "screen.navigation.category.order:Integer=20",
+	service = ScreenNavigationCategory.class
 )
 public class CommerceTaxMethodFixedRatesScreenNavigationCategory
-	implements ScreenNavigationCategory,
-			   ScreenNavigationEntry<CommerceTaxMethod> {
+	implements ScreenNavigationCategory {
 
 	public static final String CATEGORY_KEY = "tax-rates";
 
@@ -72,16 +45,11 @@ public class CommerceTaxMethodFixedRatesScreenNavigationCategory
 	}
 
 	@Override
-	public String getEntryKey() {
-		return ENTRY_KEY;
-	}
-
-	@Override
 	public String getLabel(Locale locale) {
 		ResourceBundle resourceBundle = ResourceBundleUtil.getBundle(
 			"content.Language", locale, getClass());
 
-		return _language.get(resourceBundle, ENTRY_KEY);
+		return language.get(resourceBundle, ENTRY_KEY);
 	}
 
 	@Override
@@ -90,82 +58,7 @@ public class CommerceTaxMethodFixedRatesScreenNavigationCategory
 			SCREEN_NAVIGATION_KEY_COMMERCE_TAX_METHOD;
 	}
 
-	@Override
-	public boolean isVisible(User user, CommerceTaxMethod commerceTaxMethod) {
-		if (commerceTaxMethod == null) {
-			return false;
-		}
-
-		String engineKey = commerceTaxMethod.getEngineKey();
-
-		if (engineKey.equals("fixed-tax")) {
-			return true;
-		}
-
-		return false;
-	}
-
-	@Override
-	public void render(
-			HttpServletRequest httpServletRequest,
-			HttpServletResponse httpServletResponse)
-		throws IOException {
-
-		RenderRequest renderRequest =
-			(RenderRequest)httpServletRequest.getAttribute(
-				JavaConstants.JAVAX_PORTLET_REQUEST);
-
-		CommerceTaxFixedRatesDisplayContext
-			commerceTaxFixedRatesDisplayContext =
-				new CommerceTaxFixedRatesDisplayContext(
-					_commerceChannelLocalService,
-					_commerceChannelModelResourcePermission,
-					_commerceCurrencyLocalService, _commerceTaxFixedRateService,
-					_commerceTaxMethodService, _cpTaxCategoryService,
-					_percentageFormatter, renderRequest);
-
-		httpServletRequest.setAttribute(
-			WebKeys.PORTLET_DISPLAY_CONTEXT,
-			commerceTaxFixedRatesDisplayContext);
-
-		_jspRenderer.renderJSP(
-			_servletContext, httpServletRequest, httpServletResponse,
-			"/tax_rates.jsp");
-	}
-
 	@Reference
-	private CommerceChannelLocalService _commerceChannelLocalService;
-
-	@Reference(
-		target = "(model.class.name=com.liferay.commerce.product.model.CommerceChannel)"
-	)
-	private ModelResourcePermission<CommerceChannel>
-		_commerceChannelModelResourcePermission;
-
-	@Reference
-	private CommerceCurrencyLocalService _commerceCurrencyLocalService;
-
-	@Reference
-	private CommerceTaxFixedRateService _commerceTaxFixedRateService;
-
-	@Reference
-	private CommerceTaxMethodService _commerceTaxMethodService;
-
-	@Reference
-	private CPTaxCategoryService _cpTaxCategoryService;
-
-	@Reference
-	private JSPRenderer _jspRenderer;
-
-	@Reference
-	private Language _language;
-
-	@Reference
-	private PercentageFormatter _percentageFormatter;
-
-	@Reference(
-		target = "(osgi.web.symbolicname=com.liferay.commerce.tax.engine.fixed.web)"
-	)
-	private ServletContext _servletContext;
+	protected Language language;
 
 }
